fix: move Apps badge subscription into AppsDialog via smart/view split
Repo-owner review on PR #404: having NavigationBar subscribe to
AppsService and pass a refreshed `apps` array down to AppsDialog created
an unwanted dependency between NavigationBar and the Apps-settings
feature. AppsDialog should own its own service subscription instead.

Splits AppsDialog into a thin smart AppsDialog.tsx (fetches apps via
appsService.openSettings(), subscribes to 'connected'/'disconnected',
no incyclist-services import in the view) and a pure AppsDialogView.tsx
carrying the original presentational logic, mirroring the live
WorkoutDetailsDialog/WorkoutDetailsView and NavigationBar/NavigationBarView
split already used in this codebase. Storybook (AppsDialog.stories.tsx)
now targets AppsDialogView directly with plain mock apps, same pattern
as NavigationBar.stories.tsx targeting NavigationBarView - no service
mocking needed.

NavigationBar.tsx reverts to rendering <AppsDialog /> without owning any
apps state or AppsService subscription itself. The subscription-behavior
test moves from NavigationBar.test.tsx to AppsDialog.test.tsx; the
former now only asserts NavigationBar renders AppsDialog without an
`apps` prop.
