+++ src/components/AppsDialog/index.ts
@@ -1,2 +1,3 @@
 export * from './AppsDialog';
+export * from './AppsDialogView';
 export * from './types';
